+++ userfun/sensor.h
@@ -2,7 +2,7 @@
 #define SENSOR_H
 #include <stdint.h>
 
-uint8_t sensor_detect(void);
+float sensor_detect(void);
 int DifPWM(void);
 
 #endif // SENSOR_H
